+++ ThrEd4/globals.cpp
@@ -30,7 +30,7 @@
 namespace fs = std::filesystem;
 
 uint32_t   ActiveColor = 0; // active color selector
-uint32_t   ActiveLayer = 0; // active layer
+uint8_t    ActiveLayer = 0; // active layer
 fRECTANGLE AllItemsRect {}; // rectangle enclosing all forms and stitches
 
 std::vector<fPOINT>* AngledFormVertices; // form formOrigin data for angle fills
